kubeflow-anomaly-detection/requirements.txt
absl-py==0.10.0
adal==1.2.4
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
astunparse==1.6.3
attrs==20.1.0
azure-common==1.1.25
azure-mgmt-storage==11.2.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0
backcall==0.2.0
black==20.8b1
bleach==3.1.5
boto3==1.14.55
botocore==1.17.55
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
configparser==5.0.0
cryptography==3.1
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
docker==4.3.1
docutils==0.15.2
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-logging==1.15.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
httplib2==0.18.1
ibm-cos-sdk==2.7.0
ibm-cos-sdk-core==2.7.0
ibm-cos-sdk-s3transfer==2.7.0
idna==2.8
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kfp==1.0.0
kfp-server-api==1.0.0
kfserving==0.4.0
kubeflow-fairing==1.0.2
kubeflow-pytorchjob==0.1.3
kubeflow-tfjob==0.1.3
kubernetes==10.0.1
Markdown==3.2.2
MarkupSafe==1.1.1
minio==6.0.0
mistune==0.8.4
msrest==0.6.18
msrestazure==0.6.4
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
strip-hints==0.1.9
table-logger==0.3.6
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.24.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
